Wrap iterator around float not ing

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -7,7 +7,7 @@ import asyncio
 from typing import Iterator
 
 
-async def async_generator() -> Iterator[int]:
+async def async_generator() -> Iterator[float]:
     """
     async_generator - Function to wait for a bit and yield a random int
     Returns: A generator
